fix(sidekick): remove sync utility CORS request
Keep the catalog-context helpers with the externally loaded sync module so opening the Sidekick does not require a second CORS-enabled asset.
